Add ParameterService implementation and rework parameter API.
This adds the parameter services implementation for rclpy. Along the way
I made some tweaks to the rclpy Parameter class in order to accomodate
actual use. The biggest change is that rather than trying to infer a
type from the python type of the value now a type must be specified from
a Python Enum that maps to the numeric interface enum.